+++ packages/moonstone/Dropdown/Dropdown.js
@@ -16,22 +16,21 @@
  * @exports DropdownBaseDecorator
  */
 
-
+import kind from '@enact/core/kind';
+import EnactPropTypes from '@enact/core/internal/prop-types';
+import {handle, forward} from '@enact/core/handle';
 import Changeable from '@enact/ui/Changeable';
 import Toggleable from '@enact/ui/Toggleable';
-import equals from 'ramda/src/equals';
-import EnactPropTypes from '@enact/core/internal/prop-types';
 import Group from '@enact/ui/Group';
-import kind from '@enact/core/kind';
 import Pure from '@enact/ui/internal/Pure';
 import PropTypes from 'prop-types';
-import compose from 'ramda/src/compose';
+import {compose, equals} from 'ramda';
 import React from 'react';
 
-import {Button} from '../Button/Button';
+import Button from '../Button';
 import ContextualPopupDecorator from '../ContextualPopupDecorator/ContextualPopupDecorator';
-import {Item} from '../Item/Item';
-import {Scroller} from '../Scroller/Scroller';
+import Item from '../Item';
+import Scroller from '../Scroller';
 import Skinnable from '../Skinnable';
 
 import css from './Dropdown.module.less';
@@ -103,28 +102,39 @@ const DropdownList = Skinnable(
 			 *
 			 * @type {Number}
 			 */
-			selected: PropTypes.number
+			selected: PropTypes.number,
+
+			/**
+			 * The width of DropdownList.
+			 *
+			 * @type {('large'|'medium'|'small')}
+			 */
+			width: PropTypes.oneOf(['large', 'medium', 'small'])
 		},
 
 		styles: {
-			className: 'dropDownList',
-			css
+			css,
+			className: 'dropDownList'
+		},
+
+		computed: {
+			className: ({width, styler}) => styler.append(width)
 		},
 
 		render: ({children, onSelect, selected, ...rest}) => {
+			delete rest.width;
+
 			return (
-				<div {...rest}>
-					<Group
-						childComponent={Item}
-						className={css.group}
-						component={children ? Scroller : null}
-						onSelect={onSelect}
-						select="radio"
-						selected={selected}
-					>
-						{children}
-					</Group>
-				</div>
+				<Group
+					{...rest}
+					childComponent={Item}
+					component={children ? Scroller : null}
+					onSelect={onSelect}
+					select="radio"
+					selected={selected}
+				>
+					{children}
+				</Group>
 			);
 		}
 	})
@@ -220,24 +230,29 @@ const DropdownBase = kind({
 		 * @required
 		 * @public
 		 */
-		title: PropTypes.string
+		title: PropTypes.string,
+
+		/**
+		 * The width of Dropdown.
+		 *
+		 * @type {('large'|'medium'|'small')}
+		 * @default 'medium'
+		 * @public
+		 */
+		width: PropTypes.oneOf(['large', 'medium', 'small'])
 	},
 
 	defaultProps: {
 		direction: 'down',
-		open: false
+		open: false,
+		width: 'medium'
 	},
 
 	handlers: {
-		onSelect: (ev, {onClose, onSelect}) => {
-			if (onClose) {
-				onClose(ev);
-			}
-
-			if (onSelect) {
-				onSelect(ev);
-			}
-		}
+		onSelect: handle(
+			forward('onSelect'),
+			forward('onClose')
+		)
 	},
 
 	styles: {
@@ -246,6 +261,7 @@ const DropdownBase = kind({
 	},
 
 	computed: {
+		className: ({width, styler}) => styler.append(width),
 		title: ({children, selected, title}) => {
 			const isSelectedValid = !(typeof selected === 'undefined' || selected === null || selected >= children.length || selected < 0);
 
@@ -258,12 +274,13 @@ const DropdownBase = kind({
 		}
 	},
 
-	render: ({children, disabled, onOpen, onSelect, open, selected, title, ...rest}) => {
-		const popupProps = {children, onSelect, open, selected};
+	render: ({children, disabled, onOpen, onSelect, open, selected, width, title, ...rest}) => {
+		const popupProps = {children, onSelect, selected, width};
 
 		// `ui/Group`/`ui/Repeater` will throw an error if empty so we disable the Dropdown and prevent Dropdown to open if there are no children.
 		const hasChildren = children && children.length;
 		const openDropdown = hasChildren ? open : false;
+		delete rest.width;
 
 		return (
 			<ContextualButton