+++ packages/sampler/stories/default/Dropdown.js
@@ -30,6 +30,7 @@ storiesOf('Moonstone', module)
 					onSelect={action('onSelect')}
 					size={select('size', ['small', 'large'], Config)}
 					title={text('title', Config, 'Dropdown')}
+					width={select('width', ['small', 'medium', 'large'], Config)}
 				>
 					{items}
 				</Dropdown>